+++ CHANGELOG.md
@@ -7,6 +7,7 @@
  * No longer show help for sub-commands by default 
  * Warnings about invalid accounts/roles in config.yaml are now Debug messages #980
  * Default ProfileFormat is now the `Friendly` format #992
+ * `config`, `config-profiles` and `completions` are now sub-commands of `setup` #975
 
 ### New Features